Change `versionNumber`'s type back to `uint256`

+++ modules/contracts/contracts/adjudicator/libs/LibStateChannelApp.sol
@@ -32,7 +32,7 @@ contract LibStateChannelApp {
     struct AppChallenge {
         address latestSubmitter;
         bytes32 appStateHash;
-        uint128 versionNumber;
+        uint256 versionNumber;
         uint248 finalizesAt;
         ChallengeStatus status;
     }
